Fix uninstall CoreMetrics (3x)

+++ Shell/MainService.cs
@@ -1444,7 +1444,10 @@ namespace Bhp.Shell
                 return true;
             }
             var pluginName = args[1];
-            Directory.Delete(Path.Combine("Plugins", pluginName), true);
+            if (Directory.Exists(Path.Combine("Plugins", pluginName)))
+            {
+                Directory.Delete(Path.Combine("Plugins", pluginName), true);
+            }
             File.Delete(Path.Combine("Plugins", $"{pluginName}.dll"));
             Console.WriteLine($"Uninstall successful, please restart bhp-cli.");
             return true;
